Move list_code_definition_names to a tool file (#2097)
